backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
google-genai==1.0.0
pydantic==2.9.0
python-dotenv==1.0.1
motor==3.5.1
pymongo==4.8.0
